requirements.txt
requests
beautifulsoup4
pandas
scikit-learn
numpy
matplotlib
pillow
openai
python-dotenv
scikit-learn
ultralytics
